observability/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Kevin James <KevinJames@thekev.in>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
prometheus-client = "^0.11.0"
python-json-logger = "v2.0.0"
opentelemetry-api = "^1.0.0"
opentelemetry-sdk = "^1.0.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
